fix: MISRA C:2012 Rule 10.3 compliance - narrowing conversions (Issue #845)
This commit fixes all ~106 MISRA 10.3 violations that were flagging
narrowing conversions and cross-type-category assignments without
explicit casts.

Key changes:

1. BitUtils.ts: Added narrow type casts for bit write operations
   - multiBitWrite, singleBitWrite wrap RHS with cast for u8/u16/i8/i16
   - writeOnlySingleBit, writeOnlyMultiBit similarly updated

2. UnaryExprGenerator.ts: Handle -2147483648 and -9223372036854775808
   - These literals overflow in C (parsed as -(value_too_big_for_int))
   - Now generates (int32_t)INT32_MIN and (int64_t)INT64_MIN instead
   - Automatically includes <limits.h> when these are used

3. CodeGenerator.ts: Cast limit macros in float-to-int clamp expressions
   - INT32_MAX/INT32_MIN have type 'int', not 'int32_t'
   - Now casts to target type: (int32_t)INT32_MAX, etc.

4. VariableDeclHelper.ts: Cross-type-category casts for variable init
   - Int-to-float: adds (float) or (double) cast
   - Float-to-int: adds target type cast for MISRA compliance

5. SimpleHandler.ts: Int-to-float casts for assignment statements
   - Same logic as variable declarations but for assignments

6. TypeResolver.ts: Type inference for unsuffixed literals
   - Plain integers now return "int" type
   - Plain floats now return "f64" type
   - Enables proper cross-category detection

7. NarrowingCastHelper.ts: Added cross-type-category detection
   - isFloatType(), isIntegerType() helpers
   - isCrossTypeCategoryConversion() for detection
   - wrapIntToFloat() for int-to-float casting

MISRA 10.3 violations: 106 -> 0 (100% fixed)
All 941 integration tests pass
All 5452 unit tests pass

+++ tests/bit-indexing/bit-single-write.test.cpp
@@ -11,13 +11,13 @@
 // Tests: setting individual bits in integer
 int main(void) {
     uint8_t flags = 0U;
-    flags = (flags & ~(1U << 0)) | (1U << 0);
-    flags = (flags & ~(1U << 3)) | (1U << 3);
-    flags = (flags & ~(1U << 7)) | (1U << 7);
+    flags = (uint8_t)((flags & ~(1U << 0)) | (1U << 0));
+    flags = (uint8_t)((flags & ~(1U << 3)) | (1U << 3));
+    flags = (uint8_t)((flags & ~(1U << 7)) | (1U << 7));
     if (flags != 0x89) {
         return 1;
     }
-    flags = (flags & ~(1U << 3)) | (0U << 3);
+    flags = (uint8_t)((flags & ~(1U << 3)) | (0U << 3));
     if (flags == 0x81) {
         return 0;
     }
